Added Simulated Annealing with Restarts
Simulated Annealing has the inherent problem that its configuration
depends on both the budget (in function evaluations) and the range of
the objective function.

One idea to mitigate the dependency on the budget is to internally
perform independent restarts with increasing budgets.
This is implemented in the new "sars" file set.
The algorithm implemented there is equivalent to the Simulated
Annnealing variant given in the file set "sa", with the only difference
being that it will start an internal run with a budget of 1024 FEs and
doubles the budget in every restart.
The internal run is then configured based on the internal budget exactly
in the same way that the original Simulated Annealing algorithm is
configured based on the whole budget.

In other words, sars wastes about half of the budget in the worst case
for internal runs that are too short, but in return does not suffer the
problem that all improvements happen closely to the end of the overall
budget.

The configuration will still be sub-optimal, because the range of the
objective function values are not known. These are necessary to
determine proper start and end temperatures. Since we use the problem
dimension instead of the information on the objective function, we
cannot really expect very good performance.
But at least the algorithm is now not also additionally necessarily
slow.

The other files (opoea, opofea, sa) have been adapted to use the exactly
same initial solutions as sars.